LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.get.double-expected.txt

2d.imageData.get.double

createImageData(w, h) double is converted to long


PASS createImageData(w, h) double is converted to long
